Encapsulate timeline & hierarchy into `moxie/ui/`
